pyproject.toml
[project]
name = "crop-yield-prediction-pakistan"
version = "0.1.0"
description = "Data collection and preprocessing pipeline for crop yield prediction in Pakistan"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.0",
    "pymongo>=4.8.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.0",
    "jupyter>=1.1.0",
    "ruff>=0.6.0",
]

[tool.uv]
package = false
